1.7 Added Remove-InBoxApps and Cleaned up params
Updated the way mandatory params are handled. I hate having to type out the command line params when I run this, so I like to use defaults that meet my needs. I changed mandatory params to use [ValidateNotNullOrEmpty()] instead so it will still prompt if you missed a param.  I also added a function to remove In-Box apps using an XML list as a source.